Added queen moves generation. Fixed issue with comparison of byte and 256

+++ ChessProgram/QueemCore/BitBoard/BitBoardSerializer.cs
@@ -20,9 +20,9 @@ namespace QueemCore
 			{
 				squares[rank] = new List<Square>[256];
 				
-				for (byte b = 0; b < 256; ++b)
+				for (int b = 0; b < 256; ++b)
 				{
-					squares[rank][b] = GetList(rank, b);
+					squares[rank][b] = GetList(rank, (byte)b);
 				}
 			}
